Refactor Bar chart pt2
