+++ Sources/SwiftUICharts/Common/Colors.swift
@@ -1,5 +1,5 @@
 //
-//  Color+Ext.swift
+//  Colors.swift
 //  SwiftUICharts
 //
 //  Created by András Samu on 2019. 07. 19..
@@ -75,20 +75,4 @@ extension Color {
         }
         self.init(red: Double(r) / 255, green: Double(g) / 255, blue: Double(b) / 255)
     }
-}
-
-class HapticFeedback {
-    #if os(watchOS)
-    // watchOS implementation
-    static func playSelection() {
-        WKInterfaceDevice.current().play(.click)
-    }
-
-    #else
-    // iOS implementation
-    let selectionFeedbackGenerator = UISelectionFeedbackGenerator()
-    static func playSelection() {
-        UISelectionFeedbackGenerator().selectionChanged()
-    }
-    #endif
 }
